requirements.txt
torch
librosa
numpy
scipy
matplotlib
pandas
tqdm
pyyaml
python-dotenv
fastapi
uvicorn
